Refactor Logging API (#230)
