+++ Source/AwsCommonRuntimeKit/crt/Logger.swift
@@ -1,0 +1,84 @@
+//  Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+//  SPDX-License-Identifier: Apache-2.0.
+
+import AwsCCommon
+var logger: aws_logger?
+
+public struct Logger {
+    public static func initilize(pipe: UnsafeMutablePointer<FILE>?, level: LogLevel) {
+        // Clean up the logger if it was previously initialized
+        if var logger = logger {
+            aws_logger_clean_up(&logger)
+            aws_logger_set(nil)
+        }
+        logger = aws_logger()
+        var options = aws_logger_standard_options()
+        options.level = level.rawValue
+        options.file = pipe
+        aws_logger_init_standard(&logger!, allocator.rawValue, &options)
+        aws_logger_set(&logger!)
+    }
+
+    public static func initilize(filePath: String, level: LogLevel) {
+        // Clean up the logger if it was previously initialized
+        if var logger = logger {
+            aws_logger_clean_up(&logger)
+            aws_logger_set(nil)
+        }
+        logger = aws_logger()
+        
+        filePath.withCString { cFilePath in
+            var options = aws_logger_standard_options()
+            options.level = level.rawValue
+            options.filename = cFilePath
+            aws_logger_init_standard(&logger!, allocator.rawValue, &options)
+            aws_logger_set(&logger!)
+        }
+    }
+}
+
+public enum LogLevel {
+    case none
+    case fatal
+    case error
+    case warn
+    case info
+    case debug
+    case trace
+}
+
+extension LogLevel {
+
+    public static func fromString(string: String) -> LogLevel {
+        switch string {
+        case "TRACE":
+            return .trace
+        case "INFO":
+            return .info
+        case "WARN":
+            return .warn
+        case "DEBUG":
+            return .debug
+        case "FATAL":
+            return .fatal
+        case "ERROR":
+            return .error
+        case "NONE":
+            return .none
+        default:
+            return .none
+        }
+    }
+
+    var rawValue: aws_log_level {
+        switch self {
+        case .none: return AWS_LL_NONE
+        case .fatal: return AWS_LL_FATAL
+        case .error: return AWS_LL_ERROR
+        case .warn: return AWS_LL_WARN
+        case .info: return AWS_LL_INFO
+        case .debug: return AWS_LL_DEBUG
+        case .trace: return AWS_LL_TRACE
+        }
+    }
+}
